add: make Checkbox from legacy DarkCheckbox, LightCheckbox
