+++ src/components/ui/Form/DarkCheckbox/index.tsx
@@ -1,21 +1,6 @@
-import type { InputHTMLAttributes } from 'react';
-import { useId } from 'react';
-import InputLayout from '../InputLayout'
-import { CheckboxDisplayAreaCss, CheckboxAreaShellCss, CheckboxDivCss, CheckboxLabelDefaultCss, CheckboxLabelFromTextCss, CheckboxTagDefaultCss } from './style';
+import type { CheckboxProps } from '../Checkbox';
+import Checkbox from '../Checkbox';
 
-export default function DarkCheckbox({
-  className, ...props
-}: InputHTMLAttributes<HTMLInputElement> & CssProps) {
-  const inputId = useId()
-  const { ...inputProps } = props
-  return <InputLayout css={CheckboxDivCss} className={className}>
-    <InputLayout.Shell css={CheckboxAreaShellCss}>
-      <div css={CheckboxDisplayAreaCss}>
-        <InputLayout.Tag css={[CheckboxTagDefaultCss]} {...inputProps} id={inputId} type="checkbox" />
-      </div>
-      <InputLayout.Label css={[CheckboxLabelDefaultCss, CheckboxLabelFromTextCss]} htmlFor={inputId}>
-        {props.placeholder}
-      </InputLayout.Label>
-    </InputLayout.Shell>
-  </InputLayout>
-}
+const LegacyDarkCheckbox = ({ placeholder, ...props }: CheckboxProps) => <Checkbox.dark {...props} label={placeholder} />
+
+export default LegacyDarkCheckbox
